requirements.txt
python-dotenv==1.0.1

requests~=2.28.2
telebot~=0.0.5
pyTelegramBotAPI~=4.14.0